Update prob_calculator.py
updated the draw method:
if the number of balls drawn is larger than than the amount of balls in the hat, the results should be all balls.
if the number of balls drawn exceedes the number of the contents, or the hats contents is zero. return all the balls back into the hat